# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Api Documentation",
        "description": "L5 Swagger",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0."
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/api/v1/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register User",
                "operationId": "328da17a0e06913b4ee6b37e84b9ef1d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/post/store": {
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "Create Post",
                "operationId": "ef9cec93a2dbec1450abae499d803b4a",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "title": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "example name",
                                    "title": "example title",
                                    "description": "example description"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "name"
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "title"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "description"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "fail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/post/get/{id}": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Get Detail post",
                "operationId": "8f0ad629822a94693ad35dd6778b2e66",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "name"
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "title"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "description"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/post/update/{id}": {
            "put": {
                "tags": [
                    "Post"
                ],
                "summary": "Update Post",
                "operationId": "9a788f07e11e0bc83d7d2147e36a08ad",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "title": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "example name",
                                    "title": "example title",
                                    "description": "example description"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "name"
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "title"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "description"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/post/delete/{id}": {
            "delete": {
                "tags": [
                    "Post"
                ],
                "summary": "Delete Post",
                "operationId": "298a4a98f2b483e6ef21dc286d2515a1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "delete todo success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/post/gets": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Get List Post",
                "operationId": "b7ac27f555c6b47a8b891a7e7b5a79ea",
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "_id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "example name"
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "example title"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "example description"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2021-12-11T09:25:53.000000Z"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2021-12-11T09:25:53.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CreateUserRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "anhnt"
                    },
                    "email": {
                        "type": "string",
                        "example": "anhnguyen@gmail.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "123456"
                    }
                },
                "type": "object"
            },
            "GetUserRequest": {
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "anhnguyen@gmail.com"
                    }
                },
                "type": "object"
            },
            "AuthResource": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": "true"
                    },
                    "data": {
                        "properties": {
                            "token": {
                                "type": "string",
                                "example": "3|PXX3pewsSBbtJhJQXXuRZ1NiLHzTvD6Bv2TBUJjm"
                            },
                            "token_type": {
                                "type": "string",
                                "example": "Bearer"
                            }
                        },
                        "type": "object"
                    },
                    "message": {
                        "type": "string",
                        "example": "success"
                    }
                },
                "type": "object"
            },
            "LoginResource": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": "true"
                    },
                    "data": {
                        "properties": {
                            "token": {
                                "type": "string",
                                "example": "3|PXX3pewsSBbtJhJQXXuRZ1NiLHzTvD6Bv2TBUJjm"
                            },
                            "token_type": {
                                "type": "string",
                                "example": "Bearer"
                            }
                        },
                        "type": "object"
                    },
                    "message": {
                        "type": "string",
                        "example": "success"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format (Bearer )",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}